+++ src/main/java/edu/fscj/cen3024c/simplecalculator/SimpleCalculator.java
@@ -1,0 +1,33 @@
+//SimpleCalculator.java
+//L. Nguyen
+//1/22/2024
+//Creating methods for each operation
+
+package edu.fscj.cen3024c.simplecalculator;
+
+public class SimpleCalculator
+{
+    //Method for Multiplying
+    public static int multiply(int number1, int number2)
+    {
+        return number1 * number2;
+    }
+
+    //Method for dividing
+    public static int divide( int number1, int number2)
+    {
+        return number1/number2;
+    }
+    //Method for adding
+    public static int add(int number1, int number2)
+    {
+        return number1 + number2;
+    }
+
+    //Method for subtracting
+    public static int subtract(int number1, int number2)
+    {
+        return number1 - number2;
+    }
+
+}
